+++ BrokenTC2/assets/update/changelog.md
@@ -4,5 +4,4 @@
 <br/><br/>
 
 # Bugs patched
-- Last controller selectionned should now be restored on startup (second try)
-- Fixed update mechanism
+- One button per controller was not recognized